Changed Text field to use antd. Made own directory
