+++ web/components/chat/ChatTextField/ChatTextField.module.scss
@@ -1,0 +1,4 @@
+.root {
+    height: 2rem;
+    color: var(--black);
+}
